+++ src/main/java/study/tdd/simpleboard/config/JwtSecurityConfig.java
@@ -7,16 +7,19 @@ import org.springframework.security.web.DefaultSecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import study.tdd.simpleboard.api.member.entity.MemberRepository;
 import study.tdd.simpleboard.filter.JWTFilter;
-import study.tdd.simpleboard.util.JWTTokenProvider;
+import study.tdd.simpleboard.util.JWTProvider;
 
+/**
+ * JWT Security Filter 를 등록하는 Config 클래스 입니다.
+ */
 @RequiredArgsConstructor
 public class JwtSecurityConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-    private final JWTTokenProvider jwtTokenProvider;
+    private final JWTProvider jwtProvider;
     private final MemberRepository memberRepository;
 
     @Override
-    public void configure(HttpSecurity builder) throws Exception {
-        JWTFilter jwtFilter = new JWTFilter(jwtTokenProvider, memberRepository);
+    public void configure(HttpSecurity builder) {
+        JWTFilter jwtFilter = new JWTFilter(jwtProvider, memberRepository);
 
         builder.addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class);
     }
