+++ src/main/java/study/tdd/simpleboard/custom/UserCustom.java
@@ -2,11 +2,15 @@ package study.tdd.simpleboard.custom;
 
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
 import study.tdd.simpleboard.api.member.entity.Member;
 
-import java.util.*;
+import java.util.Collection;
 
+/**
+ *
+ * Authentication 내부에 들어갈 객체
+ *
+ */
 public class UserCustom extends User {
 
 //    private final Long memberId;
